requirements.txt
keras==3.1.1
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
pygame==2.5.2
scikit_learn==1.2.2
tensorflow==2.16.1
tensorflow_intel==2.16.1
